share/src/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cvzone==1.5.6
cycler==0.11.0
decorator==5.1.1
Django==4.1.6
django-cors-headers==3.13.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
filelock==3.10.7
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.1.0
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.6.3
mediapipe==0.9.1.0
ml-dtypes==0.0.4
mpmath==1.3.0
networkx==3.0
ninja==1.11.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tifffile==2023.3.21
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.2.3
whitenoise==6.4.0
wrapt==1.14.1
zipp==3.15.0
